requirements.txt
asgiref==3.8.1
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.68.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pydantic==1.8.2
pytest==8.2.2
pytest-asyncio==0.23.7
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.15.0
